% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-obj.R
\name{eq_constr}
\alias{eq_constr}
\title{Make an equality constraint object to be used by in \code{sip} to create a \code{sipo}.}
\usage{
eq_constr(fn, gr.x = NULL, value)
}
\arguments{
\item{fn}{Equality function \code{g(x,t)}.}

\item{value}{Value for equality.}

\item{fn.bounds}{Lower and/or upper bound of \code{fn} input argument \code{x}. One, but not both, may be infinite.}
}
\value{
A constraint object (\code{class = c("constraint","equality","finite")})
}
\description{
Only one constraint may be specified per object.
}
