% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-obj.R
\name{finite_equality_constraint}
\alias{finite_equality_constraint}
\title{Make an equality constraint object to be used by in \code{sip} to create a \code{sipo}.}
\usage{
finite_equality_constraint(fun, grad.x = NULL, value)
}
\arguments{
\item{fun}{Equality function \code{g(x,t)}.}

\item{value}{Value for equality.}

\item{fun.bounds}{Lower and/or upper bound of \code{fun} input argument \code{x}. One, but not both, may be infinite.}
}
\value{
A constraint object (\code{class = c("constraint","equality","finite")})
}
\description{
Only one constraint may be specified per object.
}
