% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-obj.R
\name{finite_inequality_constraint}
\alias{finite_inequality_constraint}
\title{Make a (finite) inequality constraint object to be used by in \code{sip} to create a \code{sipo}.}
\usage{
finite_inequality_constraint(fun, grad.x = NULL, fun.bounds)
}
\arguments{
\item{fun}{Inequality function \code{g(x,t)}.}

\item{fun.bounds}{Lower and/or upper bound of \code{fun} input argument \code{x}. One, but not both, may be infinite.}
}
\value{
A constraint object (\code{class = c("constraint","inequality","finite")}).
}
\description{
Only one infinite constraint may be specified per object.
}
