% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-obj.R
\name{ineq_constr}
\alias{ineq_constr}
\title{Make a (finite) inequality constraint object to be used by in \code{sip} to create a \code{sipo}.}
\usage{
ineq_constr(fn, gr.x = NULL, fn.bounds)
}
\arguments{
\item{fn}{Inequality function \code{g(x,t)}.}

\item{fn.bounds}{Lower and/or upper bound of \code{fn} input argument \code{x}. One, but not both, may be infinite.}
}
\value{
A constraint object (\code{class = c("constraint","inequality","finite")}).
}
\description{
Only one infinite constraint may be specified per object.
}
