% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-obj.R
\name{inf_ineq_constr}
\alias{inf_ineq_constr}
\title{Make a infinite inequality constraint object to be used by in \code{sip} to create a \code{sipo}.}
\usage{
inf_ineq_constr(fn, gr.x = NULL, gr.t = NULL, fn.bounds, t.bounds,
  t.start = NULL)
}
\arguments{
\item{fn}{Inequality function \code{g(x,t)}.}

\item{gr.x}{(Optional) gradient (vector) of inequality constraint function w.r.t \code{x}.}

\item{gr.t}{(Optional) gradient (vector) of inequality constraint function w.r.t \code{t}.}

\item{fn.bounds}{Lower and/or upper bound of \code{fn} input argument \code{x}. One, but not both, may be infinite.}

\item{t.bounds}{List of compact index set(s) for \code{t} in which inequality constraint \code{fn} must be satisfied. Must be finite.}

\item{t.start}{(Optional) where to start the optimisation of the lower level problem. Defaults to \code{mean(ineq.tset)}.}
}
\value{
An infinite constraint object (\code{class = c("constraint","inequality","infinite")})
}
\description{
Only one infinite constraint may be specified per object.
}
