% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint-obj.R
\name{infinite_inequality_constraint}
\alias{infinite_inequality_constraint}
\title{Make a infinite inequality constraint object to be used by in \code{sip} to create a \code{sipo}.}
\usage{
infinite_inequality_constraint(fun, grad.x = NULL, grad.t = NULL,
  fun.bounds, t.bounds, t.start = NULL)
}
\arguments{
\item{fun}{Inequality function \code{g(x,t)}.}

\item{grad.x}{(Optional) gradient (vector) of inequality constraint function w.r.t \code{x}.}

\item{grad.t}{(Optional) gradient (vector) of inequality constraint function w.r.t \code{t}.}

\item{fun.bounds}{Lower and/or upper bound of \code{fun} input argument \code{x}. One, but not both, may be infinite.}

\item{t.bounds}{List of compact index set(s) for \code{t} in which inequality constraint \code{fun} must be satisfied. Must be finite.}

\item{t.start}{(Optional) where to start the optimisation of the lower level problem. Defaults to \code{mean(ineq.tset)}.}
}
\value{
An infinite constraint object (\code{class = c("constraint","inequality","infinite")})
}
\description{
Only one infinite constraint may be specified per object.
}
