% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim-exact-penalty-method.R
\name{optim_sip}
\alias{optim_sip}
\title{optim() with exact penalty method for semi-infinite programming.}
\usage{
optim_sip()
}
\arguments{
\item{pars}{starting parameter vector.}

\item{func}{Function to optimise}

\item{grad}{Gradient of func}

\item{lower}{Lower bounds for x}

\item{upper}{Upper bounds for x}

\item{fin_eq}{(finite) inequality constraint(s) as list of function(s).}

\item{fin_ineq}{(finite) equality constraint(s) as list of function(s)}

\item{inf_ineq}{Infinite inequality constraint(s) as list of function(s)}

\item{inf_ineq_index}{Set for inf_ineq to vary across}

\item{control}{list of control arguments}
}
\description{
Based on paper: Q Lin, R Loxton, KL Teo, YH Wu, C Yu (2014) http://dx.doi.org/10.1016/j.cam.2013.11.010.
}
