% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsolnp-sip.R
\name{optim_solnp}
\alias{optim_solnp}
\title{Rsolnp::solnp for semi-infinite programming.}
\usage{
optim_solnp(pars, sip, control = list(), ...)
}
\arguments{
\item{pars}{starting parameter vector.}

\item{sip}{SIP object created by \code{\link{make_sip}}.}

\item{control}{See control argument in \code{\link[Rsolnp]{solnp}}.}
}
\value{
Results from \code{\link[Rsolnp]{solnp}} as sipo object (see \code{\link{sip}}).
}
\description{
Wrapper for \code{\link[Rsolnp]{solnp}} to solve semi-infinite program. \code{solnp} uses the augmented Lagrange method for non-linear optimisation.
}
