% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-constraint.R
\name{print.constraint}
\alias{print.constraint}
\title{Methods for class "constraint"}
\usage{
\method{print}{constraint}(cns, padding = F, ...)
}
\arguments{
\item{cns}{The constraint object.}

\item{padding}{Add LHS padding to output.}

\item{...}{Additional arguments to be passed to function}
}
\description{
Methods for class "constraint"
}
