% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-sipo.R
\name{print.sipo}
\alias{print.sipo}
\title{Methods for class "sipo"}
\usage{
\method{print}{sipo}(sip, ...)
}
\arguments{
\item{sip}{The sipo object.}

\item{...}{Additional arguments to be passed to function}
}
\description{
Methods for class "sipo"
}
