% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip-obj.R
\name{sip}
\alias{sip}
\title{Make a semi-infinite program R object, \code{sipo}.}
\usage{
sip(fn, gr = NULL, lower = NULL, upper = NULL, constraint,
  x.feasible = NULL, ...)
}
\arguments{
\item{fn}{Objective function to be minimized.}

\item{gr}{(Optional) gradient (vector) of objective function \code{fn}.}

\item{lower}{(Optional) Lower bound of \code{x}.}

\item{upper}{(Optional) Upper bound of \code{x}.}

\item{constraint}{List of constraint objects to be used.}

\item{x.feasible}{(Optional) List of feasible x (vector) points for optimisation.}

\item{...}{(Optional) Other items to store in the sipo.}
}
\value{
A sipo: semi-infinite programming object
}
\description{
Makes a \code{sipo} object to be optimised by SIP solvers included in \code{sipr}. \code{sipo} object contains objective function, constraints, and known feasible points (if applicable).
}
