% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip-problem-set.R
\name{sip_hsphere_covers_hcube}
\alias{sip_hsphere_covers_hcube}
\title{Problem set for SIP problems.}
\usage{
sip_hsphere_covers_hcube(cube.centre, cube.diameter)
}
\arguments{
\item{cube.centre}{Centre of cube.}

\item{cube.diameter}{Diameter of cube.}
}
\value{
A predefined SIP problem.
}
\description{
Create SIP objects \code{sipo} for given  SIP problems.
}
\details{
Set covering: hyper-sphere covering hyper-cube
}
