% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-exact-penalty.R
\name{sip_penalty}
\alias{sip_penalty}
\title{Make exact penalty for semi-infinite programming.}
\usage{
sip_penalty(inf_ineq, inf_ineq_index)
}
\arguments{
\item{inf_ineq}{Infinite inequality constraint(s) as list of function(s).}

\item{inf_ineq_index}{List of ranges infinite inequality (inf_ineq) to vary across.}
}
\description{
Based on paper: Q Lin, R Loxton, KL Teo, YH Wu, C Yu (2014) http://dx.doi.org/10.1016/j.cam.2013.11.010.
}
