% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip-problem-set.R
\name{sip_problem}
\alias{sip_problem}
\title{Problem set from XXX for SIP problems.}
\usage{
sip_problem(number)
}
\arguments{
\item{number}{Problem number}
}
\value{
A predefined SIP problem
}
\description{
Create SIP objects \code{sipo} from classis SIP problems.
}
\details{
Problem 1: Original problem
Problem 2: # Ex. 3 Sect. 4 JJ Ruckman & A Shapiro (2009)
}
