% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sip-obj.R
\name{sip}
\alias{sip}
\title{Make a semi-infinite program R object, \code{sipo}.}
\usage{
sip(obj.fn, obj.gr = NULL, x.bounds = NULL, eq.fn = NULL, eq.gr = NULL,
  eq.val = NULL, ineq.fn, ineq.bounds, ineq.tset, ineq.gr.x = NULL,
  ineq.gr.t = NULL, x.feasible, ...)
}
\arguments{
\item{obj.fn}{Objective function to be minimized.}

\item{obj.gr}{(Optional) gradient (vector) of objective function \code{obj.fn}.}

\item{x.bounds}{(Optional) Lower and upper bound of \code{fn} input argument \code{x}. As list, or 2-row matrix if multiple.}

\item{eq.fn}{(Optional) Equality constraint (vector) function.}

\item{eq.gr}{(Optional) gradient (vector) of equality constraint function.}

\item{eq.val}{(Optional) Value(s) equality constraint(s) are equal to.}

\item{ineq.fn}{Inequality constraint function. First argument for objective input variable \code{x}. Second argument for index set variable \code{t}.}

\item{ineq.bounds}{Lower and upper bound of \code{fn} input argument \code{x}. One, but not both, may be infinite.}

\item{ineq.tset}{Compact index set for \code{t} in which inequality constraint \code{ineq.fn} must be satisfied. Must be finite.}

\item{ineq.gr.x}{(Optional) gradient (vector) of inequality constraint function w.r.t \code{x}.}

\item{ineq.gr.t}{(Optional) gradient of inequality constraint function w.r.t \code{t}.}

\item{x.feasible}{List (or one) of feasible \code{x}.}

\item{...}{(Optional) Other items to store in the sipo.}

\item{t.start}{(Optional) where to start the optimisation of the lower level problem. Defaults to \code{mean(ineq.tset)}.}
}
\value{
A sipo: semi-infinite programming object
}
\description{
Makes a \code{sipo} object to be optimised by SIP solvers included in \code{sipr}. \code{sipo} object contains objective function, constraints, and known feasible points (if applicable).
}
